fix(): socios

+++ models/Articles.php
@@ -347,8 +347,8 @@ class Articles extends \yii\db\ActiveRecord
      */
     public static function isOwner()
     {
-        $model = Partners::find()->where(['user_id' => (int) (User::id())]);
-        $id = $model->exists() ? (int) ($model->one()->id) : '';
+        $model = Partners::find()->where(['user_id' => User::id()]);
+        $id = $model->exists() ? $model->one()->id : '';
         return self::find()->where(['partner_id' => $id])->exists();
     }
